Feat: Add express error handling and fix frontend tests

+++ catAPI/src/routes/images.ts
@@ -1,26 +1,27 @@
 import * as express from 'express';
-import { Request, Response } from 'express';
+import { Request, Response, NextFunction } from 'express';
 import { Image } from '../initDB';
-// import { passportSetup } from '../auth/index';
 const jwt = require('jsonwebtoken');
 import * as dotenv from 'dotenv';
-
-// passportSetup();
 
 dotenv.config();
 
 const router = express.Router();
 
-export const authOptions = (req: Request, res: Response, next: any) => {
+export const authOptions = (
+    req: Request,
+    res: Response,
+    next: NextFunction
+) => {
     try {
         const authHeader = req.headers.authorization;
         const token = authHeader?.split(' ')[1];
-        jwt.verify(token, process.env.PRIVATE_KEY);
+        const decoded = jwt.verify(token, process.env.PRIVATE_KEY);
+        req.user = decoded;
         next();
     } catch (error: unknown) {
         console.error(error);
-        res.status(401);
-        res.send(JSON.stringify(error));
+        res.status(401).send(JSON.stringify(error));
     }
 };
 
@@ -29,7 +30,7 @@ router.get(
     async (req: Request, res: Response): Promise<void> => {
         try {
             const all = await Image.findAll();
-            res.send(all);
+            res.json(all);
         } catch (error: unknown) {
             console.error(error);
             res.send(JSON.stringify(error));
@@ -43,7 +44,13 @@ router.get(
         try {
             const id = req.params.id;
             const image = await Image.findByPk(id);
-            res.send(image);
+            if (image === null) {
+                res.status(404).send(
+                    JSON.stringify('Could not find that image')
+                );
+            } else {
+                res.json(image);
+            }
         } catch (error: unknown) {
             console.error(error);
             res.send(JSON.stringify(error));
@@ -79,23 +86,28 @@ router.delete(
     authOptions,
     async (req: Request, res: Response): Promise<void> => {
         const id = req.params.id;
-
         try {
             const image = await Image.findByPk(id);
-            const name = image !== null && image.get('title');
-            // admin users can delete any images, but regular users
-            // can only delete their own images
-            if (req.user?.admin || image?.get('userId') === req.user?.id) {
-                await Image.destroy({
-                    where: {
-                        id,
-                    },
-                });
-                res.send(`Deleted ${name} successfully`);
-            } else {
-                res.send(
-                    "Can't delete: You can only delete others' images if you are an admin."
+            if (image === null) {
+                res.status(404).send(
+                    JSON.stringify('Could not find that image')
                 );
+            } else {
+                const name = image !== null && image.get('title');
+                // admin users can delete any images, but regular users
+                // can only delete their own images
+                if (req.user?.admin || image?.get('userId') === req.user?.id) {
+                    await Image.destroy({
+                        where: {
+                            id,
+                        },
+                    });
+                    res.send(`Deleted ${name} successfully`);
+                } else {
+                    res.send(
+                        "Can't delete: You can only delete others' images if you are an admin."
+                    );
+                }
             }
         } catch (error: unknown) {
             console.log(error);
@@ -114,25 +126,31 @@ router.put(
         const { title, url, userId } = req.body;
         try {
             const image = await Image.findByPk(id);
-            // admin users can edit any images, but regular users
-            // can only edit their own images
-            if (
-                req.user?.admin ||
-                image?.get('userId') === String(req.user?.id)
-            ) {
-                await Image.update(
-                    {
-                        title,
-                        url,
-                        userId,
-                    },
-                    { where: { id } }
+            if (image === null) {
+                res.status(404).send(
+                    JSON.stringify('Could not find that image')
                 );
-                res.send('Edited image successfully');
             } else {
-                res.send(
-                    "Can't edit: You can only edit others' images if you are an admin."
-                );
+                // admin users can edit any images, but regular users
+                // can only edit their own images
+                if (
+                    req.user?.admin ||
+                    image?.get('userId') === String(req.user?.id)
+                ) {
+                    await Image.update(
+                        {
+                            title,
+                            url,
+                            userId,
+                        },
+                        { where: { id } }
+                    );
+                    res.send('Edited image successfully');
+                } else {
+                    res.send(
+                        "Can't edit: You can only edit others' images if you are an admin."
+                    );
+                }
             }
         } catch (error: unknown) {
             console.error(error);
